refactor: move skill and command patches to managers
