+++ JotunnLib/Managers/CommandManager.cs
@@ -2,11 +2,13 @@
 using System.Collections.ObjectModel;
 using Jotunn.Entities;
 using Jotunn.ConsoleCommands;
+using System.Text.RegularExpressions;
+using System.Linq;
 
 namespace Jotunn.Managers
 {
     /// <summary>
-    ///     Handles loading of all custom console and chat commands.
+    ///     Manager for handling custom console and chat commands.
     /// </summary>
     public class CommandManager : IManager
     {
@@ -62,6 +64,8 @@ namespace Jotunn.Managers
         /// </summary>
         public void Init()
         {
+            On.Console.InputText += Console_InputText;
+
             AddConsoleCommand(new HelpCommand());
             AddConsoleCommand(new ClearCommand());
         }
@@ -87,6 +91,58 @@ namespace Jotunn.Managers
             }
 
             _consoleCommands.Add(cmd);
+        }
+
+        private void Console_InputText(On.Console.orig_InputText orig, Console self)
+        {
+            orig(self);
+
+            string text = self.m_input.text;
+            string[] parts = text.Split(' ');
+
+            if (string.IsNullOrEmpty(text) && parts.Length == 0)
+            {
+                self.Print("Invalid command");
+                return;
+            }
+
+            ConsoleCommand cmd = CommandManager.Instance.ConsoleCommands.FirstOrDefault(c => c.Name == parts[0]);
+
+            // If we found a command, execute it
+            if (cmd != null)
+            {
+                // Prioritizing quoted strings, then all strings of non-white chars 
+                string[] args = Regex.Matches(text, @"""[^""]+""|\S+")
+                    .Cast<Match>()
+                    // get rid of the quotes around arguments
+                    .Select(x => x.Value.Trim('"'))
+                    // we don't need the command itself here
+                    .Skip(1)
+                    .ToArray();
+
+                cmd.Run(args);
+                return;
+            }
+
+            // If a default command, don't display error
+            if (CommandManager.DefaultConsoleCommands.Contains(parts[0]))
+            {
+                return;
+            }
+
+            // If a cheat command, check if cheats enabled
+            if (CommandManager.DefaultCheatConsoleCommands.Contains(parts[0]))
+            {
+                if (!Console.instance.IsCheatsEnabled())
+                {
+                    self.Print("Cannot use this command without cheats enabled. Use 'devcommands' to enable cheats");
+                }
+
+                return;
+            }
+
+            // Display error otherwise
+            self.Print("Invalid command");
         }
     }
 }
